Defines the initial testing framework with Catch
Pelanet *should be a library*. This way, there should be no executable
binary created when one runs `make`. Instead, the goal is to generate
the `pelanet.a` file, which is then statically linkable to any program
using it.

To run the program (and check that whatever is implemented is fine),
the idea is to use Catch (a testing framework).

 Changes to be committed:
	modified:   ../Layer.h
	modified:   ../Makefile
	deleted:    ../main.cpp
	new file:   dummy_test.cpp
	new file:   main.cpp